Update install_mqtt.sh
add missing checks from install.sh

+++ src/install_mqtt.sh
@@ -29,6 +29,39 @@ FILES="RingBuffer.mpy \
 
 MAIN_FILE="asynchio5.py"
 
+# create my_secrets.py if it does not exist. Since RedRover does not 
+# require WiFi credentials, we can provide default values.
+if [ ! -f my_secrets.py ]; then
+    echo "Creating my_secrets.py file. Please edit it with your WiFi credentials."
+    cat > my_secrets.py <<EOL
+# my_secrets.py for RedRover
+PASS=None
+SSID="RedRover"
+MQTT_SERVER="pepper.physics.cornell.edu"
+EOL
+fi
+
+# compile the RingBuffer module - creates mpy file
+mpy-cross RingBuffer.py
+
+# check for missing files in MAIN_FILE and FILES
+MISSING_FILES=0
+for f in $FILES; do
+  if [ ! -f $f ]; then
+    echo "File $f not found!"
+    MISSING_FILES=1
+  fi
+done
+if [ ! -f $MAIN_FILE ]; then
+    echo "Main file $MAIN_FILE not found!"
+    MISSING_FILES=1
+fi
+if [ $MISSING_FILES -ne 0 ]; then
+    echo "One or more files are missing. Aborting."
+    exit 1
+fi
+
+
 # uninstall / nuke the flash fs
 mpremote fs rm -r :
 mpremote fs tree -h
